Renamed exceptions.h to rascsi_exceptions.h
