+++ src/raspberrypi/controllers/scsi_controller.cpp
@@ -16,7 +16,7 @@
 
 #include "log.h"
 #include "gpiobus.h"
-#include "exceptions.h"
+#include "rascsi_exceptions.h"
 #include "devices/scsi_host_bridge.h"
 #include "devices/scsi_daynaport.h"
 #include "scsi_controller.h"
