+++ src/raspberrypi/devices/device.cpp
@@ -11,7 +11,7 @@
 #include "rascsi_version.h"
 #include "os.h"
 #include "log.h"
-#include "exceptions.h"
+#include "rascsi_exceptions.h"
 #include "device.h"
 
 unordered_set<Device *> Device::devices;
