+++ src/raspberrypi/devices/disk.cpp
@@ -17,7 +17,7 @@
 #include "os.h"
 #include "fileio.h"
 #include "device_factory.h"
-#include "exceptions.h"
+#include "rascsi_exceptions.h"
 #include "disk.h"
 #include "mode_page_device.h"
