+++ src/raspberrypi/devices/scsi_command_util.cpp
@@ -8,7 +8,7 @@
 //---------------------------------------------------------------------------
 
 #include "log.h"
-#include "exceptions.h"
+#include "rascsi_exceptions.h"
 #include "scsi_command_util.h"
 
 using namespace scsi_defs;
