+++ src/raspberrypi/rascsi_exceptions.h
