+++ src/raspberrypi/test/device_factory_test.cpp
@@ -8,7 +8,7 @@
 //---------------------------------------------------------------------------
 
 #include "testing.h"
-#include "exceptions.h"
+#include "rascsi_exceptions.h"
 #include "rascsi_version.h"
 #include "devices/device.h"
 #include "devices/device_factory.h"
