+++ src/raspberrypi/test/device_test.cpp
@@ -8,7 +8,7 @@
 //---------------------------------------------------------------------------
 
 #include "testing.h"
-#include "exceptions.h"
+#include "rascsi_exceptions.h"
 #include "devices/device.h"
 
 class TestDevice : public Device
