+++ src/raspberrypi/test/mode_pages_test.cpp
@@ -9,7 +9,7 @@
 
 #include "testing.h"
 #include "spdlog/spdlog.h"
-#include "exceptions.h"
+#include "rascsi_exceptions.h"
 #include "devices/scsi_command_util.h"
 #include "devices/scsihd.h"
 #include "devices/scsihd_nec.h"
